// src/interfaces/cat.interface.ts
export interface ICatInterface {
    created_at: string
    height: number
    id: string
    original_filename: string
    sub_id: string
    url: string
    width: number
    votes: number
    isFavourite: boolean
    favoriteId: string
    disableVoting: boolean
    disableFavorite: boolean
    disableDelete: boolean
}